Revert "feat: parser"
This reverts commit 16a373fc6574a419fa1fe046d612317b9e14995b.
