+++ src/app.module.ts
@@ -16,7 +16,6 @@ import { DateFormatModule } from './dateTimeFormat/dateFormat.module';
 import { Deal, Trade } from './deals/entities/deals.entity';
 import { Report } from './report/entities/report.entity';
 import { MailModule } from './mail/mail.module';
-import { TradeModule } from './trade/trade.module';
 
 @Module({
   imports: [
@@ -39,7 +38,6 @@ import { TradeModule } from './trade/trade.module';
       inject: [ConfigService],
     }),
     ReportModule,
-    TradeModule,
     CurrencyRateModule,
     NormalizeTradesModule,
     NormalizeReportsModule,
