+++ src/normalizeReports/normalizeReports.module.ts
@@ -1,13 +1,8 @@
 import { Module } from '@nestjs/common';
 import { NormalizeReportsService } from './normalizeReports.service';
 import { NormalizeTradesService } from 'src/normalizeTrades/normalizeTrades.service';
-import { ReportReaderService } from 'src/reportReader/reportReader.service';
 
 @Module({
-  providers: [
-    NormalizeReportsService,
-    NormalizeTradesService,
-    ReportReaderService,
-  ],
+  providers: [NormalizeReportsService, NormalizeTradesService],
 })
 export class NormalizeReportsModule {}
