+++ src/report/report.module.ts
@@ -14,7 +14,6 @@ import { DateTimeFormatService } from 'src/dateTimeFormat/dateFormat.service';
 import { DealsModule } from 'src/deals/deals.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Report } from './entities/report.entity';
-import { TradeModule } from 'src/trade/trade.module';
 
 @Module({
   imports: [
@@ -25,8 +24,6 @@ import { TradeModule } from 'src/trade/trade.module';
     ReportReaderModule,
     DateFormatModule,
     DealsModule,
-    ReportModule,
-    TradeModule,
   ],
   exports: [TypeOrmModule],
   controllers: [ReportController],
