+++ src/report/types/interfaces/report.interface.ts
@@ -1,16 +1,4 @@
-interface AccounAtStart {
-  ticker: string;
-  quantity: number;
-}
-
-export type AccounAtStartType = Record<
-  AccounAtStart['ticker'],
-  AccounAtStart['quantity']
->;
-
 export interface Report<T> {
   dateStart: string;
   trades: T[];
-  accountAtStart?: AccounAtStartType;
-  accountAtEnd?: AccounAtStartType;
 }
