+++ src/report/types/interfaces/trade.interface.ts
@@ -9,5 +9,3 @@ export interface Trade {
   ticker: string;
   currency: string;
 }
-
-export type GroupedTrades = Record<Trade['ticker'], Trade[]>;
