--- src/trade/trade.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TradeService } from './trade.service';
-
-@Module({
-  imports: [],
-  exports: [TradeService],
-  providers: [TradeService],
-})
-export class TradeModule {}
